refactor: extract InspectOptions and inferOptionsFromProject to separate options module
