+++ src/orchestrator/executor.ts
@@ -2,32 +2,13 @@
  * Orchestrator executor for inspection operations.
  */
 
-import type { Writable } from "node:stream";
 import { createDefaultInspectors } from "../builtin-inspectors/index.ts";
 import { summaryReporter } from "../builtin-reporters/index.ts";
 import { type DiagnosticSeverity, getOverallWorstSeverity } from "../diagnostics/index.ts";
 import { wrapUnexpectedExceptionsAsync } from "../error.ts";
-import { type Inspector, runInspectors } from "../inspector/index.ts";
-import type { Reporter } from "../reporter/index.ts";
-import {
-	inferParseSourceFilesOptions,
-	type ParseSourceFilesOptions,
-	parseSourceFiles,
-} from "../source-file/index.ts";
-import { parseConfig, resolveProjectPath } from "./tsconfig/index.ts";
-
-/**
- * Configuration options for inspection operations.
- */
-export interface InspectOptions {
-	sourceFilesOptions?: ParseSourceFilesOptions;
-	// biome-ignore lint/suspicious/noExplicitAny: We can't use the unknown type here because this should accept inspectors with variadic types.
-	inspectors?: Inspector<any>[];
-	/** Reporter function for formatting output (defaults to summaryReporter) */
-	reporter?: Reporter;
-	/** Output stream for writing results (defaults to process.stdout) */
-	output?: Writable;
-}
+import { runInspectors } from "../inspector/index.ts";
+import { parseSourceFiles } from "../source-file/index.ts";
+import { type InspectOptions, inferOptionsFromProject } from "./options.ts";
 
 /**
  * Executes inspection on the provided file paths with the given options.
@@ -63,16 +44,11 @@ async function inspectProjectInternal(
 	projectPath?: string,
 	options?: InspectOptions,
 ): Promise<DiagnosticSeverity | null> {
-	const tsconfigPath = await resolveProjectPath(projectPath);
-	const tsconfig = parseConfig(tsconfigPath, tsconfigPath.endsWith("jsconfig.json"));
-
-	// Create new options with computed sourceFilesOptions based on tsconfig
-	const projectOptions: InspectOptions = {
-		...options,
-		sourceFilesOptions: inferParseSourceFilesOptions(tsconfig, options?.sourceFilesOptions),
-	};
-
-	return executeInspection(tsconfig.fileNames, projectOptions);
+	const { options: projectOptions, fileNames } = await inferOptionsFromProject(
+		projectPath,
+		options,
+	);
+	return executeInspection(fileNames, projectOptions);
 }
 
 /**
